+++ src/cli/command-registry.ts
@@ -1,5 +1,6 @@
 import { parseArgs } from "./args.js";
 import type { CliOutput, CliRuntime } from "./output.js";
+import { runDoctorCommand } from "./commands/master-doctor.js";
 import { runInitCommand } from "./commands/master-init.js";
 import { getMasterCommandHelp } from "./commands/master-help.js";
 import { getPlannedCommandOutput } from "./commands/planned-command.js";
@@ -32,6 +33,8 @@ export async function runCommand(
       return 0;
     case "master-init":
       return runInitCommand(command.args, output, runtime);
+    case "master-doctor":
+      return runDoctorCommand(command.args, output, runtime);
     case "planned-command":
       output.stdout(getPlannedCommandOutput(command.command));
       return 0;